refactor: unify five duplicated macro files into shared/
AnnotationProcessorBase, RegistrationMacro, ToolProcessor, PromptProcessor,
and ResourceProcessor lived twice — once in jvm/macros/ emitting
Expr[FastMcpServer] and once in js/macros/ emitting Expr[JsMcpServer].
Their bodies mirrored almost exactly; the concrete server types were the
only meaningful difference. Lifting both to Expr[McpServerCore] lets a
single implementation serve both platforms.

- Move all five files to shared/src/.../macros/ targeting McpServerCore.
- RegistrationMacro's extension is now on McpServerCore; adds a new
  scanAnnotationsQuiet[T] overload (passes warnOnEmpty=false to the macro
  impl) for call sites that legitimately want a no-op on empty inputs.
- Delete the 10 per-platform duplicates.

MapToFunctionMacro stays per-platform: the JVM version is Jackson-backed
(JacksonConverter + ClassTag-driven container derivation) while the JS
version is zio-json-backed (JsonDecoder + Mirror derivation). Unifying
those requires a broader codec refactor outside this commit's scope.

No behavior change for users; macro-expanded registrations now take a
McpServerCore receiver rather than a platform-concrete one, but every
existing caller either ignores the result or has it flow through tests that
only care about the manager state. 138 JVM + 39 JS tests still pass.

1014 lines removed net.
