+++ fast-mcp-scala/shared/src/com/tjclp/fastmcp/macros/AnnotationProcessorBase.scala
@@ -3,18 +3,23 @@ package macros
 
 import scala.quoted.*
 
-import com.tjclp.fastmcp.server.JsMcpServer
+import com.tjclp.fastmcp.server.McpServerCore
 
-/** JS-target helper trait mirroring the JVM annotation processors but specialized to
-  * [[JsMcpServer]].
+/** Helper trait capturing the common boilerplate shared by the three annotation processors (Tool /
+  * Prompt / Resource). Generated registration code targets the abstract `McpServerCore` trait so a
+  * single shared implementation works on every backend.
   */
 private[macros] trait AnnotationProcessorBase:
 
+  /** Fetch the first annotation of the requested type on the supplied Symbol. */
   protected inline def findAnnotation[A: Type](using Quotes)(
       sym: quotes.reflect.Symbol
   ): Option[quotes.reflect.Term] =
     MacroUtils.extractAnnotation[A](sym)
 
+  /** Retrieve `(name, description)` from any annotation following the `(name: Option[String],
+    * description: Option[String], ...)` constructor pattern. Falls back to the method's Scaladoc.
+    */
   protected def nameAndDescription(using Quotes)(
       annot: quotes.reflect.Term,
       methodSym: quotes.reflect.Symbol
@@ -36,15 +41,19 @@ private[macros] trait AnnotationProcessorBase:
     val name = maybeName.getOrElse(methodSym.name)
     (name, maybeDesc.orElse(methodSym.docstring))
 
+  /** Build a stable method reference expression that survives inlining. */
   protected def methodRef(using Quotes)(
       owner: quotes.reflect.Symbol,
       method: quotes.reflect.Symbol
   ): Expr[Any] =
     MacroUtils.getMethodRefExpr(owner, method)
 
+  /** Execute the registration effect eagerly inside the default ZIO runtime, returning the server
+    * value so callers can inline the expression directly.
+    */
   protected def runAndReturnServer(
-      server: Expr[JsMcpServer]
-  )(registration: Expr[Any])(using Quotes): Expr[JsMcpServer] =
+      server: Expr[McpServerCore]
+  )(registration: Expr[Any])(using Quotes): Expr[McpServerCore] =
     '{
       import zio.*
       Unsafe.unsafe { implicit unsafe =>
